serialized tokenizer

+++ large_model.py
@@ -37,8 +37,8 @@ class LargeLanguageModel(object):
             print 'Adding output layer to gpu ' + str(current_gpu)
             current_gpu += 1
             rnn_output = tf.unpack(self.rnn, axis=1)
-            self.w_proj = tf.Variable(tf.zeros([self.vocab_size, self.hidden_dim]))
-            self.b_proj = tf.Variable(tf.zeros([self.vocab_size]))
+            self.w_proj = tf.Variable(tf.zeros([self.vocab_size, self.hidden_dim]),dtype=tf.float32)
+            self.b_proj = tf.Variable(tf.zeros([self.vocab_size]),dtype=tf.float32)
             self.output_seq = tf.placeholder(tf.int32, shape=([None, self.seq_len]))
             losses = []
             outputs = []
